feat: migrate and refactor Phase III enhancements to C++
Migrates and refactors Phase III enhancements (Reporting, Harvester,
Pruner) into a package-based C++ standalone application.

Changes:
- Restructured source into src/glia/core and src/glia/modules.
- Native C++ implementation of tech stack ID, selective extraction,
  and dependency-aware pruning.
- Transitioned SDD verification to empirical measurement standard.
- Integrated all modules into a unified sorrel console application.
- Added architectural documentation (DESIGN.md, RESTRICTIONS.md).
- Updated lifecycle tracking (sorrel_checkins.md, sorrel_checkouts.md).
